gukeyang333

+++ src/main/java/com/ujcmsitems/core/controller/CpictureController.java
@@ -4,14 +4,12 @@ import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.core.toolkit.StringUtils;
 import com.ujcmsitems.core.domain.Picture;
 import com.ujcmsitems.core.service.CpictureService;
-import com.ujcmsitems.core.service.impl.CpictureServiceImpl;
 import com.ujcmsitems.utils.MinIoUtil;
 import com.ujcmsitems.utils.R;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
-import io.swagger.models.auth.In;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -30,7 +28,7 @@ public class CpictureController {
 
     @ApiOperation(value = "上传图片")
     @PostMapping("/pictureUpload")
-    public String upload(MultipartFile file,String pictureName,Integer type) {
+    public String upload(MultipartFile file,String pictureName,Long type) {
         String bucketName = "carousel";
         System.out.println(bucketName);
         if (StringUtils.isBlank(bucketName)) {
@@ -67,7 +65,7 @@ public class CpictureController {
 
     @PostMapping("/replace")
     @ApiOperation("修改图片名称")
-    public String replace(long id, String imgName, String imgPath,Integer type){
+    public String replace(long id, String imgName, String imgPath,Long type){
         if (cpictureService.replace(id, imgName, imgPath, type)){
             return "修改成功";
         }else {
