+++ src/main/java/com/ujcmsitems/core/service/CpictureService.java
@@ -3,7 +3,6 @@ package com.ujcmsitems.core.service;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.service.IService;
 import com.ujcmsitems.core.domain.Picture;
-import io.swagger.models.auth.In;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
@@ -16,7 +15,7 @@ public interface CpictureService extends IService<Picture> {
      * @param imgName
      * @return
      */
-    boolean upload(MultipartFile file, String imgName , Integer type);
+    boolean upload(MultipartFile file, String imgName , Long type);
 
     /**
      * 获取图片信息
@@ -47,7 +46,7 @@ public interface CpictureService extends IService<Picture> {
      * @return
      */
 
-    boolean replace (Long id,String imgName,String imgPath,Integer type);
+    boolean replace (Long id,String imgName,String imgPath,Long type);
 
     /**
      * 批量删除
